flatlist implemented for staff list
