+++ screens/staff/StaffList.tsx
@@ -1,21 +1,53 @@
 import React, { useEffect, useState } from 'react';
-import { View, Text, SafeAreaView, VirtualizedList } from 'react-native';
+import { FlatList, StyleSheet, View, Text } from "react-native";
+import { Thumbnail, } from 'native-base';
 import StaffItem from "../staff/StaffItem";
 import StaffDirectoryService from "../../services/StaffDirectoryService";
+import { Theme, Style, HeaderStyle } from '../../Theme.style';
+import SearchBar from "../../components/SearchBar"
+import ToggleButton from "../staff/ToggleButton";
+import { HomeStackParamList } from '../../navigation/MainTabNavigator';
+import { StackNavigationProp } from '@react-navigation/stack';
+import { RouteProp } from '@react-navigation/native';
+import { TouchableOpacity } from 'react-native-gesture-handler';
 
-interface StaffList {
-    sortByName: boolean
+const style = StyleSheet.create({
+    content: {
+        ...Style.cardContainer, ...{
+            backgroundColor: Theme.colors.black,
+            padding: 16,
+        }
+    },
+    header: Style.header,
+    headerTitle: HeaderStyle.title,
+    searchBar: {
+        marginBottom: 16,
+    },
+})
+
+interface Params {
+    navigation: StackNavigationProp<HomeStackParamList>;
+    route: RouteProp<HomeStackParamList, 'EventDetailsScreen'>;
 }
-interface Staff {
-    FirstName: string
-    LastName: string
-    Position: string
-    Phone: string
-    Email: string
-    key: number
-}
-export default function StaffList({ sortByName }: StaffList): JSX.Element {
+
+export default function StaffList({ navigation }: Params): JSX.Element {
     const [staff, setStaff] = useState([]);
+    useEffect(() => {
+        navigation.setOptions({
+            headerShown: true,
+            title: 'Staff Directory',
+            headerTitleStyle: style.headerTitle,
+            headerStyle: { backgroundColor: Theme.colors.background },
+            headerLeft: function render() {
+                return <TouchableOpacity onPress={() => navigation.goBack()} style={{ display: 'flex', flexDirection: 'row', alignItems: 'center' }} >
+                    <Thumbnail square source={Theme.icons.white.back} style={{ width: 24, height: 24 }} />
+                    <Text style={{ color: 'white', fontSize: 16, transform: [{ translateX: -4 }] }}>More</Text>
+                </TouchableOpacity>
+            },
+            headerLeftContainerStyle: { left: 16 },
+            headerRight: function render() { return <View style={{ flex: 1 }} /> }
+        })
+    }, [])
     useEffect(() => {
         const loadStaff = async () => {
             console.log("Fetching data from site")
@@ -24,52 +56,27 @@ export default function StaffList({ sortByName }: StaffList): JSX.Element {
         if (staff.length === 0)
             loadStaff()
     }, [])
-    /* useEffect(() => {
-    setStaff(sort())
-    }, [sortByName]) */
-    /*const sort = () => {
-        if (sortByName) {
-            return staff.sort(function (a: any, b: any) {
-                return a.LastName > b.LastName;
-            })
-        } else {
-            return staff
-        }
-    }*/
-    const parseTelephone = (tel: string) => {
-        const telephone = tel.split(',')[0].replace(/\D/g, '')
-        const extension = tel.split(',')[1] ? tel.split(',')[1].replace(/\D/g, '') : ""
-        if (telephone && extension) return telephone + "," + extension
-        else return telephone
-    }
-    const getItem = (data: Staff, index: number) => {
-        return {
-            key: index.toString(),
-            FirstName: staff[index].FirstName,
-            LastName: staff[index].LastName,
-            Email: staff[index].Email,
-            Phone: parseTelephone(staff[index].Phone),
-            Position: staff[index].Position
-        }
 
-    }
 
-    const getItemCount = (data: any) => {
-        return staff.length;
-    }
+    const [searchText, setSearchText] = useState("");
+    const [sortByName, setSortByName] = useState(false);
 
     return (
-        <SafeAreaView>
-            <VirtualizedList
-                data={staff}
-                initialNumToRender={6}
-                renderItem={({ item }) => <StaffItem key={item.key} staff={item}></StaffItem>}
-                getItemCount={getItemCount}
-                getItem={getItem}
-                maxToRenderPerBatch={6}
-                windowSize={2}
-                removeClippedSubviews={true}
-            />
-        </SafeAreaView>
+        <FlatList
+            ListHeaderComponentStyle={style.header}
+            ListHeaderComponent={
+                <View style={style.content}>
+                    <SearchBar
+                        style={style.searchBar}
+                        searchText={searchText}
+                        handleTextChanged={(newStr) => setSearchText(newStr)}
+                        placeholderLabel="Search by name or location..."></SearchBar>
+                    <ToggleButton sortByName={sortByName} setSortByName={setSortByName} btnText_one={"By Location"} btnText_two={"By Last Name"}></ToggleButton>
+                </View>
+            }
+            data={staff}
+            renderItem={({ item }) => <StaffItem staff={item}></StaffItem>}
+            keyExtractor={(item) => item.FirstName + item.LastName}
+        />
     )
 }
